Make kernel independent of identity mapping
The guest will manage that area, so make sure only virtual memory in
kernel space is referenced.

Move GDT setup to the kernel as well, which required some reshuffling in
main. (It was only in EFI loader because I thought the GDT lives in
physical memory, but it does not)